login integration

+++ react_framework/src/authentication/Login.js
@@ -1,4 +1,4 @@
-import React, {useRef} from "react";
+import React, {useRef, useEffect, useState} from "react";
 import { Button, Col, Container, Form, FormGroup, FormLabel } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
 import { useMutation, gql } from "@apollo/client";
@@ -19,33 +19,39 @@ const Login = () => {
 
     const navigate = useNavigate();
     const [login, {data, loading, error}] = useMutation(LOGIN);
+    const [errorMessage, setErrorMessage] = useState('');
 
     const usernameRef = useRef(null);
     const passwordRef = useRef(null);
 
-    const submitLoginForm = (event) => {
+    const submitLoginForm = async (event) => {
         event.preventDefault();
+        setErrorMessage('');
 
         const username = usernameRef.current.value;
         const password = passwordRef.current.value;
 
         //mock response
-        try {
-            const {data} = login({
+        await login({
                 variables: {username, password}
-            });
-        } catch (error) {
-            console.error('Wrong username or password', error);
-        };
-        
-        const token = data['id'];
-        console.log(token);
-        localStorage.clear();
-        localStorage.setItem('user-token', token);
-        setTimeout(() => {
-            navigate('/');
-        }, 500);
+        });
     }
+
+    useEffect(() => {
+        if (data) {
+            console.log('Received data:', data);
+            const token = data['login']['user']['id'];
+            localStorage.clear();
+            localStorage.setItem('user-token', token);
+            navigate('/'); 
+        }
+        if (error) {
+            console.error('Received GraphQL error:', error);
+            setErrorMessage("Username or password is incorrect");
+
+        }
+    }, [data, error]);
+
     const createNewAccount = (event) => {
         event.preventDefault();
         navigate('/auth/signup');
@@ -67,6 +73,12 @@ const Login = () => {
                                 <FormLabel htmlFor={'login-password'}>Password</FormLabel>
                                 <input type={'password'} className="form-control" id={'login-password'} name="password" ref={passwordRef} required />
                             </FormGroup>
+                            {/* error if wrong username or password*/}
+                            {errorMessage && (
+                                <div className="error-message" style={{ color: 'red' }}>
+                                    {errorMessage}
+                                </div>
+                            )}
                            <div className="row justify-content-end">
                             <div className="col-auto">
                                 <Button type="submit" className="btn-primary mt-2" id="login-btn">Login</Button>
